requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers==2.7.0
transformers==4.35.2
torch==2.1.1
faiss-cpu==1.7.4
numpy==1.24.3
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
pdfplumber==0.10.2
python-multipart==0.0.6
huggingface-hub>=0.16.0,<1.0
google-generativeai